% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumble functions.R
\name{tumble_plot}
\alias{tumble_plot}
\title{Tumble Plot}
\usage{
tumble_plot(dat, outcome, target, moderator, axis_labels = c("x", "y"),
  plot_title = NA, legend_title = NA, plot_limits)
}
\arguments{
\item{dat}{A dataset of class data.frame.}

\item{outcome}{The dependent variable.}

\item{target}{The continuous independent variable of class numeric.}

\item{moderator}{The continuous moderator of class numeric.}

\item{axis_labels}{A list of two character objects for the x- and y-axes. Defaults to c('x','y').}

\item{plot_title}{A character object for the plot title. Defaults to NA.}

\item{legend_title}{A character object for the legend title. Defaults to NA.}

\item{plot_limits}{A list of two numeric objects for xlim and ylim. Defaults to the ranges of the independent and dependent variables.}
}
\description{
This function plots a tumble plot for probing a continuous by continuous interaction with empirical endpoints.
}
\examples{
tumble_plot()
}
\keyword{interaction}
\keyword{plot}
\keyword{tumble}
