% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumble functions.R
\name{tumble_set}
\alias{tumble_set}
\title{Tumble Set}
\usage{
tumble_set(dat, outcome, target, moderator)
}
\arguments{
\item{dat}{A dataset of class data.frame.}

\item{outcome}{The dependent variable.}

\item{target}{The continuous independent variable of class numeric.}

\item{moderator}{The continuous moderator of class numeric.}
}
\description{
This function prepares an interaction model for tumble plotting.
}
\examples{
tumble_set()
}
\keyword{interaction}
\keyword{plot}
\keyword{tumble}
